Force closing scp2 connection
There's a bug in scp2 v0.1.4 that surfaces with infinite recursion when trying to deploy more than once in the same run. The already-closed underlying connection is reused and it fails. There's a fix in scp2 v0.2.0, but apparently there is something else wrong in that configuration - at least I'm getting "connection refused", so it's easier for me to fix it from here.

Fixes https://github.com/dasuchin/grunt-ssh-deploy/issues/45